requirements.txt
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
SQLAlchemy>=2.0.28
python-dotenv==1.0.1
pytest==8.0.0
